optimised final version

+++ LockedMe.com/src/com/lockedme/service/LockedMeServiceImpl.java
@@ -14,18 +14,26 @@ public class LockedMeServiceImpl implements LockedMeService {
 	@Override
 	public FileLock createFileLock(FileLock filelock) throws FileLockException {
 		
+		if(!isValidName(filelock.getName())) {
+			throw new FileLockException("Entered name "+filelock.getName()+" is invalid");
+		}
 		return dao.createFileLock(filelock);
 	}
 
 	@Override
 	public void deleteFileLock(int id) throws FileLockException {
+		if(!isValidId(id)) {
+			throw new FileLockException("ID -> "+id+" is invalid...Id must be an integer and greater than 1000");
+		}
 		dao.deleteFileLock(id);
 		
 	}
 
 	@Override
 	public FileLock getFile(int id) throws FileLockException {
-	
+		if(!isValidId(id)) {
+			throw new FileLockException("ID -> "+id+" is invalid...Id must be an integer and greater than 1000");
+		}
 		return dao.getFile(id);
 	}
 
@@ -40,6 +48,22 @@ public class LockedMeServiceImpl implements LockedMeService {
 		
 		dao.exit();
 		
+	}
+	
+	private boolean isValidName(String name) {
+		boolean val=false;
+		if(name.trim().matches("[a-zA-Z _.]{1,15}")) {
+			val=true;
+		}
+		return val;
+	}
+	
+	private boolean isValidId(int id) {
+		boolean val=false;
+		if(id > 1000) {
+			val=true;
+		}
+		return val;
 	}
 
 }
